Move mixin types to types/ directory
